+++ packages/firebase-storage-kit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # firebase-storage-kit
 
+## 1.5.0
+
+### Minor Changes
+
+- [#37](https://github.com/marknesh/firebase-storage-kit/pull/37) [`ea99986`](https://github.com/marknesh/firebase-storage-kit/commit/ea99986ed6563136f2bbf7cd9125679ef1160290) Thanks [@marknesh](https://github.com/marknesh)! - Add optional `contentType` to `UploadOptions` for setting Firebase Storage MIME type at upload time.
+
+## 1.4.0
+
+### Minor Changes
+
+- [#32](https://github.com/marknesh/firebase-storage-kit/pull/32) [`b3fcd05`](https://github.com/marknesh/firebase-storage-kit/commit/b3fcd05181bbd735617037ded89791827cac9e74) Thanks [@marknesh](https://github.com/marknesh)! - Add React hooks via `firebase-storage-kit/react`: `useStorageManager`, `useStorageState`, `useUpload`, `useBatch`, and optional `StorageManagerProvider`. React is an optional peer dependency.
+
 ## 1.3.0
 
 ### Minor Changes